Satoshi Nakamoto's original white paper
The mother of all crypto papers to appreciate and understand. In the ashes of the 2008 financial crisis, Satoshi (it still blows my mind that we have no idea who Satoshi is - so genius if you think about the genesis of decentralization and cryptography...) released this white paper to offer an alternative, decentralized financial system that's governed by maths and code, without the need of a central bank. If you truly understand this paper, you are halfway there.